Check if file exists

+++ tools/amd_build/build_amd.py
@@ -132,17 +132,18 @@ if is_hip_clang():
         print("%s updated" % gloo_cmake_file)
 
 gloo_cmake_file = "third_party/gloo/cmake/Modules/Findrccl.cmake"
-do_write = False
-with open(gloo_cmake_file, "r") as sources:
-    lines = sources.readlines()
-newlines = [line.replace('RCCL_LIBRARY', 'RCCL_LIBRARY_PATH') for line in lines]
-if lines == newlines:
-    print("%s skipped" % gloo_cmake_file)
-else:
-    with open(gloo_cmake_file, "w") as sources:
-        for line in newlines:
-            sources.write(line)
-    print("%s updated" % gloo_cmake_file)
+if os.path.exists(gloo_cmake_file):
+    do_write = False
+    with open(gloo_cmake_file, "r") as sources:
+        lines = sources.readlines()
+    newlines = [line.replace('RCCL_LIBRARY', 'RCCL_LIBRARY_PATH') for line in lines]
+    if lines == newlines:
+        print("%s skipped" % gloo_cmake_file)
+    else:
+        with open(gloo_cmake_file, "w") as sources:
+            for line in newlines:
+                sources.write(line)
+        print("%s updated" % gloo_cmake_file)
 
 hipify_python.hipify(
     project_directory=proj_dir,
